fix: refactor names to use PascalCase
