+++ app/assets/javascripts/preferences/PreferencesMenu.tsx
