+++ app/assets/javascripts/preferences/components/MenuItem.tsx
