+++ app/assets/javascripts/preferences/index.ts
@@ -1,5 +1,8 @@
 import { toDirective } from '../components/utils';
-import { PreferencesViewWrapper, PreferencesWrapperProps } from './view';
+import {
+  PreferencesViewWrapper,
+  PreferencesWrapperProps,
+} from './PreferencesView';
 
 export const PreferencesDirective = toDirective<PreferencesWrapperProps>(
   PreferencesViewWrapper
