+++ app/assets/javascripts/preferences/models/index.ts
@@ -1,2 +1,2 @@
 export * from './preferences';
-export * from './twoFactorAuth';
+export * from './two-factor-auth';
